@@ -2,26 +2,28 @@
   "name": "my-app",
   "version": "0.1.0",
   "private": true,
+  "proxy": "http://localhost:3000/",
   "dependencies": {
-    "@material-ui/core": "^4.7.2",
-    "@material-ui/icons": "^4.5.1",
-    "@material-ui/styles": "^4.5.2",
-    "bootstrap": "^4.3.1",
+    "@material-ui/core": "^4.11.0",
+    "@material-ui/icons": "^4.9.1",
+    "@material-ui/styles": "^4.10.0",
+    "bootstrap": "^4.5.0",
     "express": "^4.17.1",
     "fs": "0.0.1-security",
-    "graphql-tag": "^2.10.1",
-    "jpeg-js": "^0.3.6",
+    "graphql-tag": "^2.10.4",
+    "jpeg-js": "^0.3.7",
     "prop-types": "^15.7.2",
-    "react": "^16.8.6",
-    "react-bootstrap": "^1.0.0-beta.9",
-    "react-dom": "^16.8.6",
-    "react-images-upload": "^1.2.7",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.0.1"
+    "react": "^16.13.1",
+    "react-bootstrap": "^1.2.2",
+    "react-dom": "^16.13.1",
+    "react-images-upload": "^1.2.8",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "3.0.1",
+    "react-table": "^7.3.2"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "start:all": "npm run build && cd server && npm start",
+    "start": "npm run build && cd server && npm start",
+    "start:dev": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
