@@ -4,10 +4,11 @@ const path = require('path');
 const cookieParser = require('cookie-parser');
 const logger = require('morgan');
 const cons = require('consolidate');
+const cors = require('cors');
 const indexRouter = require('./routes/index');
 const usersRouter = require('./routes/users');
 const app = express();
-
+app.use(cors());
 
 
 // view engine setup
@@ -22,6 +23,7 @@ app.use(cookieParser());
 
 app.use(express.static(path.join(__dirname, '../build')));
 
+//routing methods
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
 
@@ -30,15 +32,5 @@ app.use(function(req, res, next) {
   next(createError(404));
 });
 
-// error handler
-app.use(function(err, req, res, next) {
-  // set locals, only providing error in development
-  res.locals.message = err.message;
-  res.locals.error = req.app.get('env') === 'development' ? err : {};
-
-  // render the error page
-  res.status(err.status || 500);
-  res.render('error');
-});
 
 module.exports = app;
