@@ -2,7 +2,8 @@ import React, { Component } from 'react';
 import Toolbar from "@material-ui/core/Toolbar";
 import './App.css';
 import MainPage from './components/MainPage/MainPage';
-import Academics from './components/Academics/Academics';
+import Academics from './components/Academics/components/Academics';
+import Contact from './components/Contact/Contact';
 import Podcasts from './components/Extras/Podcasts';
 import Posts from './components/Extras/Posts';
 import AppBar from '@material-ui/core/AppBar';
@@ -31,22 +32,25 @@ class App extends Component {
               <Link to='/' style={{textDecoration: 'none', color: 'white'}}><Button className={classes.buttons}>Home</Button></Link>
               <Link to='/academics' style={{textDecoration: 'none', color: 'white'}}><Button className={classes.buttons}>Academics</Button></Link>
               <Link to='/feed' style={{textDecoration: 'none', color: 'white'}}><Button className={classes.buttons}>Test-Feed</Button></Link>
-              <Link to='/contact' style={{textDecoration: 'none', color: 'white'}}><Button className={classes.buttons}>Contact</Button></Link>
+              {/* <Link to='/contact' style={{textDecoration: 'none', color: 'white'}}><Button className={classes.buttons}>Contact</Button></Link> */}
             </Toolbar>
           </AppBar>
         <div>
         <Route path='/' exact render={() => 
           <MainPage /> }
         />
-        <Route path='/academics' exact render={() => 
+        <Route path='/academics' render={() => 
           <Academics /> }
         />
-        <Route path='/podcasts' exact render={() => 
+        <Route path='/podcasts' render={() => 
           <Podcasts /> }
         />
-        <Route path='/feed' exact render={() => 
+        <Route path='/feed' render={() => 
           <Posts /> }
         />
+         {/* <Route path='/contact' render={() => 
+          <Contact /> }
+        /> */}
        </div>
       </HashRouter>
     )
