@@ -1,14 +1,13 @@
-import React, { Component } from 'react';
-import { withStyles, Fab } from "@material-ui/core"
+import React, { useState } from 'react';
+import { makeStyles, Fab } from "@material-ui/core"
 import FavoriteIcon from "@material-ui/icons/Favorite"
-import PropTypes from 'prop-types';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import '../../App.css';
-import Mendoza from "../MainPage/Mendoza.jpg"
-import Denver from "../MainPage/Denver.jpg"
 
+const Mendoza = 'https://jela-website.s3.us-east-2.amazonaws.com/Mendoza.jpg'
+const Denver = 'https://jela-website.s3.us-east-2.amazonaws.com/Denver.jpg'
 
-const styles = theme => ({
+const useStyles = makeStyles(theme => ({
     main: {
         backgroundColor: "#d7e3f5",
         margin:'auto',
@@ -59,53 +58,39 @@ const styles = theme => ({
         fontLamily: 'PT Sans',
         fontWeight: "bold"
       }
-})
+}));
 
-class Feed extends Component {
-    constructor(props){
-        super(props);
-    }
-
-    render() {
-        const { classes } = this.props;
-        const username = this.props.username;
-        const avatar = this.props.avatar;
-        const image = this.props.image;
-        const caption = this.props.caption;
-        return (
-            <div className={classes.main}>
-                <article className={classes.post}>
-                    <header>
-                        <div className={classes.user}>
-                            <div className={classes.userAvatar}>
-                                <img className={classes.avatar} src={avatar ? avatar: Denver} alt="posted"/>
-                            </div>
-                            <div className={classes.username}>
-                                <span>{username ? username: "jlaprade"}</span>
-                            </div>
+const Feed = ({username, avatar, image, caption}) => {
+    const classes = useStyles();
+      
+    return (
+        <div className={classes.main}>
+            <article className={classes.post}>
+                <header>
+                    <div className={classes.user}>
+                        <div className={classes.userAvatar}>
+                            <img className={classes.avatar} src={avatar ? avatar: Denver} alt="posted"/>
                         </div>
-                    </header>
-                    <div>
-                        <img className={classes.postImage} src={image ? image: Mendoza} alt="pic"/>
+                        <div className={classes.username}>
+                            <span>{username ? username: "jlaprade"}</span>
+                        </div>
                     </div>
-                   
-                    <div className={classes.caption}>
-                        <Fab aria-label="like">
-                            <FavoriteIcon />
-                        </Fab>
-                        <strong className={classes.name}>
-                            {username ? username: "jlaprade"}
-                        </strong>
-                        {caption ? caption: ' aint not rest for the wicked'}
-                    </div>
+                </header>
+                <div>
+                    <img className={classes.postImage} src={image ? image: Mendoza} alt="pic"/>
+                </div>
+                <div className={classes.caption}>
+                    <Fab aria-label="like">
+                        <FavoriteIcon />
+                    </Fab>
+                    <strong className={classes.name}>
+                        {username ? username: "jlaprade"}
+                    </strong>
+                    {caption ? caption: ' aint not rest for the wicked'}
+                </div>
             </article>
         </div>
-        )
-    }
+    )
 };
 
-Feed.propTypes = {
-    classes: PropTypes.object.isRequired,
-};
-
-export default withStyles(styles)(Feed);
+export default Feed;
